moved parsers to parsing folder. Refactored displacement memory calculation method
