+++ 001-instruction-decoding-on-the-8086/Decoder/Parsing/ByteParser.cs
@@ -1,7 +1,7 @@
 using System;
 using System.IO;
 
-namespace Homework001;
+namespace Homework001.Parsing;
 
 public static class ByteParser
 {
@@ -17,17 +17,20 @@ public static class ByteParser
 
     public static string DisplacementMemoryModeImplementation(byte r_m, short displacementValue)
     {
+        bool hasDisplacement = displacementValue != 0;
+        int absoluteDisplacementValue = Math.Abs(displacementValue);
         string sign = displacementValue < 0 ? "-" : "+";
+
         return r_m switch
         {
-            0b000 => displacementValue == 0 ? $"[bx + si]" : $"[bx + si {sign} {Math.Abs(displacementValue)}]",
-            0b001 => displacementValue == 0 ? $"[bx + di]" : $"[bx + di {sign} {Math.Abs(displacementValue)}]",
-            0b010 => displacementValue == 0 ? $"[bp + si]" : $"[bp + si {sign} {Math.Abs(displacementValue)}]",
-            0b011 => displacementValue == 0 ? $"[bp + di]" : $"[bp + di {sign} {Math.Abs(displacementValue)}]",
-            0b100 => displacementValue == 0 ? $"[si]" : $"[si {sign} {Math.Abs(displacementValue)}]",
-            0b101 => displacementValue == 0 ? $"[di]" : $"[di {sign} {Math.Abs(displacementValue)}]",
-            0b110 => displacementValue == 0 ? $"[bp]" : $"[bp {sign} {Math.Abs(displacementValue)}]",
-            0b111 => displacementValue == 0 ? $"[bx]" : $"[bx {sign} {Math.Abs(displacementValue)}]",
+            0b000 => $"[bx + si{(hasDisplacement ? $" {sign} {absoluteDisplacementValue}" : "")}]",
+            0b001 => $"[bx + di{(hasDisplacement ? $" {sign} {absoluteDisplacementValue}" : "")}]",
+            0b010 => $"[bp + si{(hasDisplacement ? $" {sign} {absoluteDisplacementValue}" : "")}]",
+            0b011 => $"[bp + di{(hasDisplacement ? $" {sign} {absoluteDisplacementValue}" : "")}]",
+            0b100 => $"[si{(hasDisplacement ? $" {sign} {absoluteDisplacementValue}" : "")}]",
+            0b101 => $"[di{(hasDisplacement ? $" {sign} {absoluteDisplacementValue}" : "")}]",
+            0b110 => $"[bp{(hasDisplacement ? $" {sign} {absoluteDisplacementValue}" : "")}]",
+            0b111 => $"[bx{(hasDisplacement ? $" {sign} {absoluteDisplacementValue}" : "")}]",
             _ => throw new InvalidOperationException("invalid reg value")
         };
     }
